#1485 refactor: introduce `ObjectConditionsTrait`
