+++ plugins/BEdita/Core/src/Model/Action/BaseAction.php
@@ -73,23 +73,4 @@ abstract class BaseAction
     {
         return $this->execute($data);
     }
-
-    /**
-     * Allowed object `status` condition
-     *
-     * @return array Empty array if all `status` are allowed otherwise a list of allowed values
-     */
-    protected function statusCondition()
-    {
-        $filter = [
-            'on' => ['status' => 'on'],
-            'draft' => ['status IN' => ['on', 'draft']],
-        ];
-        $level = Configure::read('Status.level');
-        if ($level && isset($filter[$level])) {
-            return $filter[$level];
-        }
-
-        return [];
-    }
 }
